backend/requirements.txt
firebase_admin
uvicorn
FastAPI
python-dotenv

